+++ src/main/java/timelessodyssey/model/game/elements/Tile.java
@@ -1,7 +1,7 @@
 package timelessodyssey.model.game.elements;
 
 public class Tile extends Element {
-    public Tile(int x, int y) {
+    public Tile(double x, double y) {
         super(x, y);
     }
 }
