+++ src/main/java/timelessodyssey/states/GameState.java
@@ -29,7 +29,7 @@ public class GameState extends State<Scene> {
     }
 
     @Override
-    public void step(Game game, GUI gui, long time) throws IOException, URISyntaxException, FontFormatException {
+    public void step(Game game, GUI gui, double time) throws IOException, URISyntaxException, FontFormatException {
         GUI.Action action = gui.getNextAction();
         controller.step(game, action, time);
         screenViewer.draw(gui);
